pyproject.toml
[project]
name = "ai-agent-org"
version = "0.1.0"
description = "Multi-agent AI organization: PM → Engineer → Reviewer pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.50",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff>=0.11",
    "bandit[toml]>=1.8",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "S", "A", "C4", "PT", "RET", "SIM"]
ignore = ["S603", "S607"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101", "B112"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 0
